refactor: consolidate metric settings drawer into single reusable component
